chore: Move DropParser to its own file
